Fix voice/sms preference changing to different slot
On a Subsidy lock enabled device, it is seen that till
the subsidy carrier sim latches to network, the other
slot sim stays in perso locked state, due to this subsidy
carrier sim always initialized first when compared to
non-subsidy sim. This is resulting in voice/sms preference
getting set to subsidy carrier sim always after boot.

Fix: If any of the sim is in perso locked state, wait for
its subId to be created before setting voice/sms preference.

Change-Id: Iab3851161f341c5261d790ef91c75d9657c798cb
CRs-Fixed: 3147491